refactor: 도메인에 맞지 않는 로직 이동
