+++ src/main/java/com/example/msglab/domain/MessageClient.java
@@ -5,5 +5,5 @@ public interface MessageClient {
     /**
      * @param message json 형태의 전송할 문자열
      */
-    MessageClientResp send(String message);
+    void send(String message);
 }
